+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added `singlePersonBooking.php` to handle when the participant is customer, contact and participant.
 - Added `__bookSingleParticipant.php` and `__bookMultipleParticipants.php` to handle different settings.
 - Fixing `frontend.js` to work with single participant-settings.
+
+## [0.9.13]
+### Added
 - Added support for attributes (customer, contact and person) in booking form.
 - Added functions to render the different attribute types, attributes not supported is multi value attributes, dates, checkbox lists and HTML
 - Added: Saving customer attributes
@@ -89,7 +92,8 @@
 - Added inquiry support in course
 
 
-[Unreleased]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.12...HEAD
+[Unreleased]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.13...HEAD
+[0.9.13]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.12...v0.9.13
 [0.9.12]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.11...v0.9.12
 [0.9.11]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.10...v0.9.11
 [0.9.10]: https://github.com/MultinetInteractive/EduAdmin-WordPress/compare/v0.9.9.2.40...v0.9.10